Backend/Packages/MAVLink_Pipeline/requirements.txt
pymavlink
RPi.GPIO
opencv-python
